Relocate `TestFind` tests from `data_test.go` to `find_test.go` for better test organization and readability. No changes to test logic.
